refactor: use UnsavedDetector instead of PromptProvider
