+++ src/locales/en/messages.po
@@ -25,11 +25,11 @@ msgstr "Accomplishments"
 
 #: src/containers/main/getMenuItems.ts:34
 #: src/pages/dashboard-page/DashboardPage.tsx:57
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:49
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:64
-#: src/pages/peer-review-page/PeerReviewPage.tsx:46
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:51
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:66
+#: src/pages/peer-review-page/PeerReviewPage.tsx:48
 #: src/pages/result-page/ResultPage.tsx:24
-#: src/pages/self-review-page/SelfReviewPage.tsx:24
+#: src/pages/self-review-page/SelfReviewPage.tsx:25
 msgid "Achievements"
 msgstr "Achievements"
 
@@ -64,11 +64,11 @@ msgstr "Are you sure you want to delete this project review?"
 
 #: src/containers/main/getMenuItems.ts:22
 #: src/pages/dashboard-page/DashboardPage.tsx:35
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:47
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:54
-#: src/pages/peer-review-page/PeerReviewPage.tsx:44
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:49
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:56
+#: src/pages/peer-review-page/PeerReviewPage.tsx:46
 #: src/pages/result-page/ResultPage.tsx:22
-#: src/pages/self-review-page/SelfReviewPage.tsx:22
+#: src/pages/self-review-page/SelfReviewPage.tsx:23
 msgid "Behavioral Competencies"
 msgstr "Behavioral Competencies"
 
@@ -76,11 +76,11 @@ msgstr "Behavioral Competencies"
 msgid "Cancel"
 msgstr "Cancel"
 
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:40
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:44
-#: src/pages/peer-review-page/PeerReviewPage.tsx:36
-#: src/pages/self-review-page/SelfReviewPage.tsx:31
-#: src/stories/decorators/promptDecorator.tsx:5
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:42
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:46
+#: src/pages/peer-review-page/PeerReviewPage.tsx:38
+#: src/pages/self-review-page/SelfReviewPage.tsx:33
+#: src/stories/decorators/promptDecorator.tsx:7
 msgid "Changes you made may not be saved."
 msgstr "Changes you made may not be saved."
 
@@ -112,11 +112,11 @@ msgstr "Delete"
 
 #: src/containers/main/getMenuItems.ts:28
 #: src/pages/dashboard-page/DashboardPage.tsx:46
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:48
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:59
-#: src/pages/peer-review-page/PeerReviewPage.tsx:45
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:50
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:61
+#: src/pages/peer-review-page/PeerReviewPage.tsx:47
 #: src/pages/result-page/ResultPage.tsx:23
-#: src/pages/self-review-page/SelfReviewPage.tsx:23
+#: src/pages/self-review-page/SelfReviewPage.tsx:24
 msgid "Dominant Characteristics"
 msgstr "Dominant Characteristics"
 
@@ -280,8 +280,8 @@ msgstr "Other people comments about your performance compared to initial expecta
 msgid "Other people comments about {name} performance compared to initial expectation"
 msgstr "Other people comments about {name} performance compared to initial expectation"
 
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:50
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:69
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:52
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:71
 msgid "Overall Evaluation"
 msgstr "Overall Evaluation"
