+++ src/locales/fa/messages.po
@@ -25,11 +25,11 @@ msgstr "دستاوردها"
 
 #: src/containers/main/getMenuItems.ts:34
 #: src/pages/dashboard-page/DashboardPage.tsx:57
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:49
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:64
-#: src/pages/peer-review-page/PeerReviewPage.tsx:46
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:51
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:66
+#: src/pages/peer-review-page/PeerReviewPage.tsx:48
 #: src/pages/result-page/ResultPage.tsx:24
-#: src/pages/self-review-page/SelfReviewPage.tsx:24
+#: src/pages/self-review-page/SelfReviewPage.tsx:25
 msgid "Achievements"
 msgstr "دستاوردها"
 
@@ -64,11 +64,11 @@ msgstr "آیا از حذف  این پروژه مطمین هستید؟"
 
 #: src/containers/main/getMenuItems.ts:22
 #: src/pages/dashboard-page/DashboardPage.tsx:35
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:47
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:54
-#: src/pages/peer-review-page/PeerReviewPage.tsx:44
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:49
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:56
+#: src/pages/peer-review-page/PeerReviewPage.tsx:46
 #: src/pages/result-page/ResultPage.tsx:22
-#: src/pages/self-review-page/SelfReviewPage.tsx:22
+#: src/pages/self-review-page/SelfReviewPage.tsx:23
 msgid "Behavioral Competencies"
 msgstr "شایستگی‌های رفتاری"
 
@@ -76,11 +76,11 @@ msgstr "شایستگی‌های رفتاری"
 msgid "Cancel"
 msgstr "لغو"
 
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:40
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:44
-#: src/pages/peer-review-page/PeerReviewPage.tsx:36
-#: src/pages/self-review-page/SelfReviewPage.tsx:31
-#: src/stories/decorators/promptDecorator.tsx:5
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:42
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:46
+#: src/pages/peer-review-page/PeerReviewPage.tsx:38
+#: src/pages/self-review-page/SelfReviewPage.tsx:33
+#: src/stories/decorators/promptDecorator.tsx:7
 msgid "Changes you made may not be saved."
 msgstr "یعنی بدون ذخیره تغییرات می‌خوای بری؟؟"
 
@@ -112,11 +112,11 @@ msgstr "حذف"
 
 #: src/containers/main/getMenuItems.ts:28
 #: src/pages/dashboard-page/DashboardPage.tsx:46
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:48
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:59
-#: src/pages/peer-review-page/PeerReviewPage.tsx:45
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:50
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:61
+#: src/pages/peer-review-page/PeerReviewPage.tsx:47
 #: src/pages/result-page/ResultPage.tsx:23
-#: src/pages/self-review-page/SelfReviewPage.tsx:23
+#: src/pages/self-review-page/SelfReviewPage.tsx:24
 msgid "Dominant Characteristics"
 msgstr "ویژگی‌های بارز"
 
@@ -280,8 +280,8 @@ msgstr "نظرات دیگران در مورد سطح عمکرد شما در ای
 msgid "Other people comments about {name} performance compared to initial expectation"
 msgstr "نظرات دیگران در مورد سطح عملکرد {name} در این پروژه"
 
-#: src/pages/manager-review-page/ManagerReviewPage.tsx:50
-#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:69
+#: src/pages/manager-review-page/ManagerReviewPage.tsx:52
+#: src/pages/manager-review-page/ManagerReviewPrintPage.tsx:71
 msgid "Overall Evaluation"
 msgstr "ارزیابی کلی"
