+++ src/pages/peer-review-page/PeerReviewPage.tsx
@@ -6,11 +6,12 @@ import { Box, Container, Divider, Paper } from '@material-ui/core';
 import { FCProps } from 'src/shared/types/FCProps';
 import { FullPageSpinner } from 'src/shared/loading';
 import { InView } from 'src/shared/in-view';
-import { PromptProvider } from 'src/shared/prompt';
 import { Redirect, Route, Switch, useParams } from 'react-router-dom';
 import { TabLink } from 'src/shared/tab';
 import { Tabs } from 'src/shared/tabs';
 import { TopStickyCard } from 'src/shared/top-sticky-card';
+import { UnsavedDetector } from 'src/shared/unsaved-detector';
+import { UnsavedPrompt } from 'src/shared/unsaved-prompt';
 import { i18n } from '@lingui/core';
 import { unescape } from 'src/shared/utils/base64.util';
 import { useLazyLoadQuery } from 'react-relay/hooks';
@@ -51,7 +52,8 @@ export default function PeerReviewPage(props: Props) {
   }
 
   return (
-    <PromptProvider message={i18n._('Changes you made may not be saved.')}>
+    <UnsavedDetector>
+      <UnsavedPrompt message={i18n._('Changes you made may not be saved.')} />
       <Container maxWidth="md">
         <InView>
           <TopStickyCard>
@@ -98,6 +100,6 @@ export default function PeerReviewPage(props: Props) {
           </Paper>
         </Box>
       </Container>
-    </PromptProvider>
+    </UnsavedDetector>
   );
 }
