--- src/shared/prompt/index.ts
@@ -1,2 +1,0 @@
-export { Prompt } from './Prompt';
-export { usePrompt, PromptProvider } from './PromptProvider';
