+++ src/shared/unload-prompt/index.ts
@@ -1,0 +1,1 @@
+export { UnloadPrompt } from './UnloadPrompt';
