+++ src/shared/unsaved-prompt/index.ts
@@ -1,0 +1,1 @@
+export { UnsavedPrompt } from './UnsavedPrompt';
